+++ catalogue/rdbms/RdbmsAdminUserCatalogue.cpp
@@ -47,7 +47,7 @@ void RdbmsAdminUserCatalogue::createAdminUser(
     if (comment.empty()) {
       throw UserSpecifiedAnEmptyStringComment("Cannot create admin user because the comment is an empty string");
     }
-    RdbmsCatalogueUtils::checkCommentOrReasonMaxLength(comment, &m_log);
+    const auto trimmedComment = RdbmsCatalogueUtils::checkCommentOrReasonMaxLength(comment, &m_log);
 
     auto conn = m_connPool->getConn();
     if (adminUserExists(conn, username)) {
@@ -84,7 +84,7 @@ void RdbmsAdminUserCatalogue::createAdminUser(
 
     stmt.bindString(":ADMIN_USER_NAME", username);
 
-    stmt.bindString(":USER_COMMENT", comment);
+    stmt.bindString(":USER_COMMENT", trimmedComment);
 
     stmt.bindString(":CREATION_LOG_USER_NAME", admin.username);
     stmt.bindString(":CREATION_LOG_HOST_NAME", admin.host);
@@ -200,7 +200,7 @@ void RdbmsAdminUserCatalogue::modifyAdminUserComment(const common::dataStructure
     if (comment.empty()) {
       throw UserSpecifiedAnEmptyStringComment("Cannot modify admin user because the comment is an empty string");
     }
-    RdbmsCatalogueUtils::checkCommentOrReasonMaxLength(comment, &m_log);
+    const auto trimmedComment = RdbmsCatalogueUtils::checkCommentOrReasonMaxLength(comment, &m_log);
 
     const time_t now = time(nullptr);
     const char *const sql =
@@ -213,7 +213,7 @@ void RdbmsAdminUserCatalogue::modifyAdminUserComment(const common::dataStructure
         "ADMIN_USER_NAME = :ADMIN_USER_NAME";
     auto conn = m_connPool->getConn();
     auto stmt = conn.createStmt(sql);
-    stmt.bindString(":USER_COMMENT", comment);
+    stmt.bindString(":USER_COMMENT", trimmedComment);
     stmt.bindString(":LAST_UPDATE_USER_NAME", admin.username);
     stmt.bindString(":LAST_UPDATE_HOST_NAME", admin.host);
     stmt.bindUint64(":LAST_UPDATE_TIME", now);