+++ catalogue/rdbms/RdbmsDiskSystemCatalogue.cpp
@@ -58,7 +58,7 @@ void RdbmsDiskSystemCatalogue::createDiskSystem(const common::dataStructures::Se
     if(comment.empty()) {
       throw UserSpecifiedAnEmptyStringComment("Cannot create disk system because the comment is an empty string");
     }
-    RdbmsCatalogueUtils::checkCommentOrReasonMaxLength(comment, &m_log);
+    const auto trimmedComment = RdbmsCatalogueUtils::checkCommentOrReasonMaxLength(comment, &m_log);
 
     auto conn = m_connPool->getConn();
     if(RdbmsCatalogueUtils::diskSystemExists(conn, name)) {
@@ -111,7 +111,7 @@ void RdbmsDiskSystemCatalogue::createDiskSystem(const common::dataStructures::Se
    stmt.bindUint64(":TARGETED_FREE_SPACE", targetedFreeSpace);
    stmt.bindUint64(":SLEEP_TIME", sleepTime);
 
-   stmt.bindString(":USER_COMMENT", comment);
+   stmt.bindString(":USER_COMMENT", trimmedComment);
 
    stmt.bindString(":CREATION_LOG_USER_NAME", admin.username);
    stmt.bindString(":CREATION_LOG_HOST_NAME", admin.host);
@@ -321,7 +321,7 @@ void RdbmsDiskSystemCatalogue::modifyDiskSystemComment(const common::dataStructu
       throw UserSpecifiedAnEmptyStringComment("Cannot modify disk system "
         "because the new comment is an empty string");
     }
-    RdbmsCatalogueUtils::checkCommentOrReasonMaxLength(comment, &m_log);
+    const auto trimmedComment = RdbmsCatalogueUtils::checkCommentOrReasonMaxLength(comment, &m_log);
 
     const time_t now = time(nullptr);
     const char *const sql =
@@ -334,7 +334,7 @@ void RdbmsDiskSystemCatalogue::modifyDiskSystemComment(const common::dataStructu
         "DISK_SYSTEM_NAME = :DISK_SYSTEM_NAME";
     auto conn = m_connPool->getConn();
     auto stmt = conn.createStmt(sql);
-    stmt.bindString(":USER_COMMENT", comment);
+    stmt.bindString(":USER_COMMENT", trimmedComment);
     stmt.bindString(":LAST_UPDATE_USER_NAME", admin.username);
     stmt.bindString(":LAST_UPDATE_HOST_NAME", admin.host);
     stmt.bindUint64(":LAST_UPDATE_TIME", now);