+++ catalogue/rdbms/RdbmsRequesterMountRuleCatalogue.cpp
@@ -74,7 +74,7 @@ void RdbmsRequesterMountRuleCatalogue::modifyRequesteMountRuleComment(
   const common::dataStructures::SecurityIdentity &admin, const std::string &instanceName,
   const std::string &requesterName, const std::string &comment) {
   try {
-    RdbmsCatalogueUtils::checkCommentOrReasonMaxLength(comment, &m_log);
+    const auto trimmedComment = RdbmsCatalogueUtils::checkCommentOrReasonMaxLength(comment, &m_log);
     const time_t now = time(nullptr);
     const char *const sql =
       "UPDATE REQUESTER_MOUNT_RULE SET "
@@ -87,7 +87,7 @@ void RdbmsRequesterMountRuleCatalogue::modifyRequesteMountRuleComment(
         "REQUESTER_NAME = :REQUESTER_NAME";
     auto conn = m_connPool->getConn();
     auto stmt = conn.createStmt(sql);
-    stmt.bindString(":USER_COMMENT", comment);
+    stmt.bindString(":USER_COMMENT", trimmedComment);
     stmt.bindString(":LAST_UPDATE_USER_NAME", admin.username);
     stmt.bindString(":LAST_UPDATE_HOST_NAME", admin.host);
     stmt.bindUint64(":LAST_UPDATE_TIME", now);
@@ -111,7 +111,7 @@ void RdbmsRequesterMountRuleCatalogue::createRequesterMountRule(const common::da
   const std::string &mountPolicyName, const std::string &diskInstanceName, const std::string &requesterName,
   const std::string &comment) {
   try {
-    RdbmsCatalogueUtils::checkCommentOrReasonMaxLength(comment, &m_log);
+    const auto trimmedComment = RdbmsCatalogueUtils::checkCommentOrReasonMaxLength(comment, &m_log);
     const auto user = User(diskInstanceName, requesterName);
     auto conn = m_connPool->getConn();
     const auto mountPolicyCatalogue = static_cast<RdbmsMountPolicyCatalogue*>(m_rdbmsCatalogue->MountPolicy().get());
@@ -168,7 +168,7 @@ void RdbmsRequesterMountRuleCatalogue::createRequesterMountRule(const common::da
     stmt.bindString(":REQUESTER_NAME", requesterName);
     stmt.bindString(":MOUNT_POLICY_NAME", mountPolicyName);
 
-    stmt.bindString(":USER_COMMENT", comment);
+    stmt.bindString(":USER_COMMENT", trimmedComment);
 
     stmt.bindString(":CREATION_LOG_USER_NAME", admin.username);
     stmt.bindString(":CREATION_LOG_HOST_NAME", admin.host);