+++ catalogue/rdbms/RdbmsTapePoolCatalogue.cpp
@@ -56,7 +56,7 @@ void RdbmsTapePoolCatalogue::createTapePool(const common::dataStructures::Securi
     if (comment.empty()) {
       throw UserSpecifiedAnEmptyStringComment("Cannot create tape pool because the comment is an empty string");
     }
-    RdbmsCatalogueUtils::checkCommentOrReasonMaxLength(comment, &m_log);
+    const auto trimmedComment = RdbmsCatalogueUtils::checkCommentOrReasonMaxLength(comment, &m_log);
 
     auto conn = m_connPool->getConn();
 
@@ -118,7 +118,7 @@ void RdbmsTapePoolCatalogue::createTapePool(const common::dataStructures::Securi
     stmt.bindBool(":IS_ENCRYPTED", encryptionValue);
     stmt.bindString(":SUPPLY", supply);
 
-    stmt.bindString(":USER_COMMENT", comment);
+    stmt.bindString(":USER_COMMENT", trimmedComment);
 
     stmt.bindString(":CREATION_LOG_USER_NAME", admin.username);
     stmt.bindString(":CREATION_LOG_HOST_NAME", admin.host);
@@ -524,7 +524,7 @@ void RdbmsTapePoolCatalogue::modifyTapePoolComment(const common::dataStructures:
     if(comment.empty()) {
       throw UserSpecifiedAnEmptyStringComment("Cannot modify tape pool because the new comment is an empty string");
     }
-    RdbmsCatalogueUtils::checkCommentOrReasonMaxLength(comment, &m_log);
+    const auto trimmedComment = RdbmsCatalogueUtils::checkCommentOrReasonMaxLength(comment, &m_log);
 
     const time_t now = time(nullptr);
     const char *const sql =
@@ -537,7 +537,7 @@ void RdbmsTapePoolCatalogue::modifyTapePoolComment(const common::dataStructures:
         "TAPE_POOL_NAME = :TAPE_POOL_NAME";
     auto conn = m_connPool->getConn();
     auto stmt = conn.createStmt(sql);
-    stmt.bindString(":USER_COMMENT", comment);
+    stmt.bindString(":USER_COMMENT", trimmedComment);
     stmt.bindString(":LAST_UPDATE_USER_NAME", admin.username);
     stmt.bindString(":LAST_UPDATE_HOST_NAME", admin.host);
     stmt.bindUint64(":LAST_UPDATE_TIME", now);